# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "FlexPrice API Service",
        "title": "FlexPrice API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/prices": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get prices with the specified filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prices"
                ],
                "summary": "Get prices",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "published",
                            "deleted",
                            "archived"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "StatusPublished",
                            "StatusDeleted",
                            "StatusArchived"
                        ],
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ListPricesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new price with the specified configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prices"
                ],
                "summary": "Create a new price",
                "parameters": [
                    {
                        "description": "Price configuration",
                        "name": "price",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePriceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.PriceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/prices/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a price by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prices"
                ],
                "summary": "Get a price by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Price ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PriceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update a price with the specified configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prices"
                ],
                "summary": "Update a price",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Price ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Price configuration",
                        "name": "price",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePriceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PriceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a price",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prices"
                ],
                "summary": "Delete a price",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Price ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreatePriceRequest": {
            "type": "object",
            "required": [
                "amount",
                "billing_cadence",
                "billing_model",
                "billing_period",
                "billing_period_count",
                "currency",
                "type"
            ],
            "properties": {
                "amount": {
                    "type": "string"
                },
                "billing_cadence": {
                    "$ref": "#/definitions/types.BillingCadence"
                },
                "billing_model": {
                    "$ref": "#/definitions/types.BillingModel"
                },
                "billing_period": {
                    "$ref": "#/definitions/types.BillingPeriod"
                },
                "billing_period_count": {
                    "type": "integer",
                    "minimum": 1
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "filter_values": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "lookup_key": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "meter_id": {
                    "type": "string"
                },
                "plan_id": {
                    "type": "string"
                },
                "tier_mode": {
                    "$ref": "#/definitions/types.BillingTier"
                },
                "tiers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreatePriceTier"
                    }
                },
                "transform_quantity": {
                    "$ref": "#/definitions/price.TransformQuantity"
                },
                "type": {
                    "$ref": "#/definitions/types.PriceType"
                }
            }
        },
        "dto.CreatePriceTier": {
            "type": "object",
            "required": [
                "unit_amount"
            ],
            "properties": {
                "flat_amount": {
                    "type": "string"
                },
                "unit_amount": {
                    "type": "string"
                },
                "up_to": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ListPricesResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "prices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PriceResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.PriceResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount stored in main currency units (e.g., dollars, not cents)\nFor USD: 12.50 means $12.50",
                    "type": "number"
                },
                "billing_cadence": {
                    "description": "BillingCadence is the billing cadence for the price ex RECURRING, ONETIME",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.BillingCadence"
                        }
                    ]
                },
                "billing_model": {
                    "description": "BillingModel is the billing model for the price ex FLAT_FEE, PACKAGE, TIERED",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.BillingModel"
                        }
                    ]
                },
                "billing_period": {
                    "description": "BillingPeriod is the billing period for the price ex month, year",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.BillingPeriod"
                        }
                    ]
                },
                "billing_period_count": {
                    "description": "BillingPeriodCount is the count of the billing period ex 1, 3, 6, 12",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "currency": {
                    "description": "Currency 3 digit ISO currency code in lowercase ex usd, eur, gbp",
                    "type": "string"
                },
                "description": {
                    "description": "Description of the price",
                    "type": "string"
                },
                "display_amount": {
                    "description": "DisplayAmount is the formatted amount with currency symbol\nFor USD: $12.50",
                    "type": "string"
                },
                "filter_values": {
                    "description": "FilterValues are the filter values for the price in case of usage based pricing",
                    "allOf": [
                        {
                            "$ref": "#/definitions/price.JSONBFilters"
                        }
                    ]
                },
                "id": {
                    "description": "ID uuid identifier for the price",
                    "type": "string"
                },
                "lookup_key": {
                    "description": "LookupKey used for looking up the price in the database",
                    "type": "string"
                },
                "metadata": {
                    "description": "Metadata is a jsonb field for additional information",
                    "allOf": [
                        {
                            "$ref": "#/definitions/price.JSONBMetadata"
                        }
                    ]
                },
                "meter_id": {
                    "description": "MeterID is the id of the meter for usage based pricing",
                    "type": "string"
                },
                "plan_id": {
                    "description": "PlanID is the id of the plan for plan based pricing",
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/types.Status"
                },
                "tenant_id": {
                    "type": "string"
                },
                "tier_mode": {
                    "description": "Tiered pricing fields when BillingModel is TIERED",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.BillingTier"
                        }
                    ]
                },
                "tiers": {
                    "description": "Tiers are the tiers for the price when BillingModel is TIERED",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/price.PriceTier"
                    }
                },
                "transform_quantity": {
                    "description": "Transform is the quantity transformation in case of PACKAGE billing model",
                    "allOf": [
                        {
                            "$ref": "#/definitions/price.JSONBTransformQuantity"
                        }
                    ]
                },
                "type": {
                    "description": "Type is the type of the price ex USAGE, FIXED",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.PriceType"
                        }
                    ]
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdatePriceRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "lookup_key": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
...
        "price.JSONBFilters": {
            "type": "object",
            "additionalProperties": {
                "type": "array",
                "items": {
                    "type": "string"
                }
            }
        },
        "price.JSONBMetadata": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "price.JSONBTransformQuantity": {
            "type": "object",
            "properties": {
                "divide_by": {
                    "description": "Divide quantity by this number",
                    "type": "integer"
                },
                "round": {
                    "description": "up or down",
                    "type": "string"
                }
            }
        },
...
        "price.PriceTier": {
            "type": "object",
            "properties": {
                "flat_amount": {
                    "description": "FlatAmount is the flat amount for the given tier and it is applied\non top of the unit amount*quantity. It solves cases in banking like 2.7% + 5c",
                    "type": "number"
                },
                "unit_amount": {
                    "description": "UnitAmount is the amount per unit for the given tier",
                    "type": "number"
                },
                "up_to": {
                    "description": "Upto is the quantity up to which this tier applies. It is null for the last tier",
                    "type": "integer"
                }
            }
        },
        "price.TransformQuantity": {
            "type": "object",
            "properties": {
                "divide_by": {
                    "description": "Divide quantity by this number",
                    "type": "integer"
                },
                "round": {
                    "description": "up or down",
                    "type": "string"
                }
            }
        },
...
        "types.BillingCadence": {
            "type": "string",
            "enum": [
                "RECURRING",
                "ONETIME"
            ],
            "x-enum-varnames": [
                "BILLING_CADENCE_RECURRING",
                "BILLING_CADENCE_ONETIME"
            ]
        },
        "types.BillingModel": {
            "type": "string",
            "enum": [
                "FLAT_FEE",
                "PACKAGE",
                "TIERED"
            ],
            "x-enum-varnames": [
                "BILLING_MODEL_FLAT_FEE",
                "BILLING_MODEL_PACKAGE",
                "BILLING_MODEL_TIERED"
            ]
        },
        "types.BillingPeriod": {
            "type": "string",
            "enum": [
                "MONTHLY",
                "ANNUAL",
                "WEEKLY",
                "DAILY"
            ],
            "x-enum-varnames": [
                "BILLING_PERIOD_MONTHLY",
                "BILLING_PERIOD_ANNUAL",
                "BILLING_PERIOD_WEEKLY",
                "BILLING_PERIOD_DAILY"
            ]
        },
        "types.BillingTier": {
            "type": "string",
            "enum": [
                "VOLUME",
                "SLAB"
            ],
            "x-enum-varnames": [
                "BILLING_TIER_VOLUME",
                "BILLING_TIER_SLAB"
            ]
        },
...
        "types.PriceType": {
            "type": "string",
            "enum": [
                "USAGE",
                "FIXED"
            ],
            "x-enum-varnames": [
                "PRICE_TYPE_USAGE",
                "PRICE_TYPE_FIXED"
            ]
        },
...
        "types.Status": {
            "type": "string",
            "enum": [
                "published",
                "deleted",
                "archived"
            ],
            "x-enum-varnames": [
                "StatusPublished",
                "StatusDeleted",
                "StatusArchived"
            ]
        },
...
        "v1.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Invalid request payload"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request payload"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter your API key in the format *x-api-key \u0026lt;api-key\u0026gt;**",
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        }
    }
}